Add an internal index-time timer to measure our work
Cleanup on exit is taking more time as structures get more complicated.
Let's not count that time.  Add an internal timer to measure only how long
it takes to build the index of words for our given file.

For reference, the current time is about 5.0s, but the times before I started
trying to add BTreeSets is only about 2.8s.  Wow -- super bad!