fix: parser accumulates numbers after they end
When a number ends we forgot to stop tracking it.  If we get a series of
numbers in a row like 123 456 789, we will calculate it as 123, 123456 and 123456789.

Fix the parser to quit tracking numbers after a word boundary.

Also fix a bug where a hex number would keep tracking even without a
leading `0x`.