Enable mouse-mode and assign actions to mousemap
I kind of hate this mode. It disables all the usual terminal selection features
and clipboard usage and forces the application to do it instead.  This would
be fine, except that putting data into the user's clipboard is super-complicated
and not universally supported across OS's.  There's some ctrl-codes to make
this work in the xterm emulator, but it may not work on Windows.
https://invisible-island.net/xterm/ctlseqs/ctlseqs.html