display: implement intraline coloring
Add intraline coloring to display lines using some canned styles referenced
via enum. Create some demo color generators that just highlight a
fixed section of text for now as a demo, or provide status line inversion
for the whole text line.

Color styles are represented as a vector of length and styles.  They
are meant to be display-width-agnostic.  It is required that the entire line
width be styled for it to be displayed.  We could use a "default" style
for any remaining text on the line, but parsing it was messy.  So I took
the easy route and just included a style for the whole line.

Implementation detail: The lengths of the style-vec may be longer than
the line itself. Any extra length is chopped off.