factor out LogFile to hold readable files
Isolate the file indexer code from the file open/reader code
by refactoring LogFile into being and giving an instance to LogFileLines.

I plan -- somehow -- to generalize the indexer into a replacement for
the filter mechanics and to make it complete lazily as needed.  Currently
it simply indexes every line, but it could be wrapped in a filter that
reduces the exposed lines further.