index: store offset to eol, not eol+1
Should we store offsets to start of line or end of line?  The other code assumes
we're indexing eol, but we were storing bol instead.  It might be nice to store
bol instead and then insert a zero for the first buffer. But this also seems
prone to confusion.

For now, fix it to record eol.  Maybe I'll change my mind later, but that's
a bridge we can cross then.