FAIL: return &str ref to line while iterating file
I cannot get the lifetimes to work out to allow me to return a &str to the line
in the file while I'm iterating the file.  Maybe it's impossible. Dunno.
Good question for the Rust community?

I tried moving the ref_next function inside the iterator struct where I
might be able to more easily specify lifetimes, but it didn't help. :-(

I plan to punt and return Strings instead.  Ick.