wip: add TextLogStream, not yet functional
Add a reader for streams which can also read from stdin, hopefully.
For now it is stuck because we can't return &[u8] to our internal buffer
because it's wrapped in a Box<> for interior mutability pattern.  So we
need to return String from our read() instead.  When I tried that earlier
it got deep into the weeds propagating the trait change to TextLogFile
and friends.  So, that's a TBD.

Meanwhile, I think this has become over-complicated.  Why not implement
Seek for Stdin instead?  All we need is a ReadBuf which never discards old
data, or something like that. Meanwhile, I've also discovered std::Cursor<>
which does most of the work.  Holy shit.  I'm rethinking all my life choices,
now. Like, the mmap has to go. Premature optimization and whatnot.