Refactor LogFile struct away; use Box<dyn ...> instead
The std library already implements Read + Seek + BufRead for Box<T>. Is it
different from ours?  Don't know, but struct TextLog is just another layer in
the mix, so replace it with just the Box itself.

We can't implement the `new` helpers for the type, without a trait, so just
make them public functions instead.

The passthrough function is weirdly more specific now, needing to call
as_ref() to get to the inner trait object.  But I expect the runtime is the same.