indexer: replace tuple struct with a named struct
Name the fields in the IndexRef with the hopes that we can
someday add a "bol" field for easier navigation.

Add a helper to construct this enum so it's less verbose and so it can
someday fill in the "bol" field.  It also, helpfully, hides the field names
we otherwise had to specify everywhere.  So, win-lose, I guess.