Remove EOL from all indexes; iterate starts only
Instead of reporting (start, end) everywhere to describe line locations, just
use `start`.  To find the line length, we can just read to \n and then trim it
the same as BufRead::Line::next().  Seems like we could be faster with
eol or len, but it's too complicated to handle all the extra noise.

Odd side-effect: we always return n+1 lines because we count the line at
the end after the last linefeed. I guess that's technically correct, but it
seems like we can filter it out by checking if start == length(). Maybe
I can fix this later.  Minor worry for now.