iterated lines return line endings
Instead of returning lines without endings and then println'ing them for cat,
return them with the line endings so cat and friends can just emit them as-is.

Works around the extra-line bug for the EOF (because line is iterated as empty string).

BODGE: This fix was actually part of some earlier squash inadvertently.
         These should probably be isolated out to a separate commit.