indexer: support iterating from an arbitrary offset in the file
This DoubleEndedIterator breaks the rust convention in that next() and next_back()
never meet.  They also don't start at the same position, so they will never duplicate
data, though.

The intended use is that the caller knows the last line he was able to iterate, and he
wants to iterate the next lines after or before that one.  So a call to iter_lines_from()
with the last already-reached line will produce this kind of iterator.  If the caller
wants to iterate backwards, he can use rev(), otherwise he will iterate forward.

Suppose I've just displayed lines 10..20 at offsets 100, 110, 120, .... 200. Now
we want to continue to the next 10 lines after these.  I can use

    iter_lines_from(200).take(10);
    # 210, 220, 230, ... 300

Or if we want to iterate the previous 10 lines (in reverse), I can use

    iter_lines_from(100).rev().take(10);
    # 90, 80, 70, ... 0