SubLineIterator: unit tests
Replicate all the existing tests for whole-line code, but customize for wrapped-lines.

Also, add some helpers for calculating expected line offsets and contents. But these
had borrow-checker problems when `file` is part of `harness` and borrowed
mutably.  Grr.  So, separate the file from the harness.  Fortunately it didn't
make the code any noisier.