Sanitize bad utf-8 input
Raw binary data is not valid utf-8.  Trying to read it causes an abort.
Sanitize it when reading it instead.

This is a stopgap for now. In the future maybe we'll go to reading
into [u8] instead of String. Then we'll be able to handle these bytes
more flexibly at display time.

Example 1:  This produces a bad file that caused an abort:

    for i in {0..255} ; do
     printf "%02x: \\$(printf '%03o' "$i")  " "$i"
     if (( (i+1)%16 == 0 )); then
       printf "\n";
     fi
    done > bad-ascii-chart.txt

Example 2: This change breaks our ability to `cat` binary files.