FilteredLog: use Box<IndexedLog> to allow recursion
We don't need a Log object in FilteredLog.  We only need IndexedLog.  But that's a trait.
It's not Sized.  But if we take a Box<IndexedLog> we can safely implement our code to
take any other IndexedLog as our inner implementation.  This is will allow us to chain
filtered logs indefinitely.  Maybe this is a bad idea.  But it's cheap enough and helps
simplify the API and reduce dependencies with Log.  So it's a Net Good, imo.

Why chaining?  It will let us have a filter for "foo" and another filter for "bar", such that if we
only want to see lines with both "foo" and "bar", we could look at
FilteredLog("foo", FilteredLog("bar", logfile)).  The inner logfile object would still be
authoritatively updated for the whole file, "bar" would also be accurate, but "foo" would be
dependent on "bar", so it would have less work to do in the end. But if "bar" went away, then
"foo" would be useless (we would have to reconstruct it).

Maybe a more sensible example is this: suppose we have a "filter-in foo"
and also a "search bar". The "filter-in" can be our Document's new "log"
source that represents the file we're showing to the user. Then the
"bar" search can be much cheaper since it ignores everything filter out
of "foo".  Since it is lazily implemented, it's also cheap to rebuild
later if we turn off the "filter-in foo".  But in the near term, it
saves us obviously wasted effort of searching lines we don't need to.

Maybe still goodness.  Maybe a crazy mess.  I can't tell yet.