Clean up some bugs around the concept of index range
Index range is complicated because the covered space is [start, end), but
the indexed lines are between [start, end].  Whether 'end' is inclusive or
not depends on if you're looking at bytes to read or lines iterated.
Furthermore, some of the code had some wrong assumptions about
the edges of indexed chunks it would see.  Sometimes I conflated the
Target and the Gap edges, too.

Clean these up and add some helper functions to make it all a bit
easier to reason about.  Hopefully.  fml.