index_chunk: always constrain loaded range to within gap
Previously we just aimed for the target-offset and let it spill over
onto the "gap" we were given.  But we actually do not need to load
any more than the gap.  And for very long lines, the target could be
far away from the gap.  Constrain it for efficiency and sanity.