faster line indexing using BufReadExt + 10MB buffer
Based on a comment about fast line parsing, I found out the BufReadExt is an auto trait
that does quick, copy-free line indexing.  It's 3x faster than before, but only for lines (not
regex) and still 2x as slow as bvr.   But this was cheap and easy, so I'm not complaining.

Will look at filter_log later using a similar technique.  I wonder if I should try to revisit
Cow<> or something else to avoid alloc/copy.