use new fast line scanner to resolve lines when seeking backwards
I had a piece of custom code to walk regions backwards and find lines, but
the new scanner is faster and cleaner.

Also: fix resolve_lines so it doesn't assume it's always reading from a whole
line start position. Only insert lines when we have both the start and end of the line.