Refactor keymap handler to support predefined sequences
A key sequence can be made of one or more events.  Define the sequence with spaces
in the key strings, like "Ctrl+X V" to represent the sequence of two keys, '^X' and 'v'.  Build
sequences to match if a partial sequence is matched.  This requires us to store the partial
sequences in the keymap so we can determine if a subsequent match is possibly going to
succeed.

Add "space" as a recognized key since " " has a different meaning.

Add extend_keymap() for testing and also to allow future expansion with user keymaps.

Add some "Esc" sequences from `less` to take advantage of this feature.  Add some tests
for matching key names, key sequences, and for the reader event handler.

TODO:
  [ ] Replace ChordKey with something that hands "-" parsing off to app
  [ ] Separate Mouse and Key events from chord consideration.
      E.g. when I press "Esc MouseWheelUp )", the  mousewheel movement should not interrupt the keyboard seq.